--- a/Gemfile
+++ b/Gemfile
@@ -102,6 +102,7 @@ gem "uglifier", "~> 4.1"
 gem "validate_url", "~> 1.0"
 gem "webpacker", "~> 3.5"
 gem "webpush", "~> 0.3"
+gem 'newrelic_rpm'
 
 group :development do
   gem "better_errors", "~> 2.5"