--- a/Procfile
+++ b/Procfile
@@ -1,3 +1,3 @@
 web: bin/start-pgbouncer bundle exec puma -C config/puma.rb
 worker: bundle exec rake jobs:work
-prometheus_exporter: bundle exec prometheus_exporter
+prometheus_exporter: bundle exec prometheus_exporter -a prometheus/custom_collector.rb