--- a/Procfile.dev
+++ b/Procfile.dev
@@ -1,3 +1,4 @@
 web: bin/rails s -p 3000
 webpacker: ./bin/webpack-dev-server
 job: bin/rake jobs:work
+prometheus_exporter: bundle exec prometheus_exporter -a prometheus/custom_collector.rb