--- a/Procfile.dev
+++ b/Procfile.dev
@@ -1,4 +1,4 @@
-web: bin/rails s -p 3000
+web: bin/rails s -p 3001
 webpacker: ./bin/webpack-dev-server
 job: bin/rake jobs:work
 prometheus_exporter: bundle exec prometheus_exporter -a prometheus/custom_collector.rb