--- a/app/views/stories/_main_stories_feed.html.erb
+++ b/app/views/stories/_main_stories_feed.html.erb
@@ -55,7 +55,9 @@
     <% if !user_signed_in? && i == 4 %>
       <%= render "stories/sign_in_invitation" %>
     <% end %>
-    <%= render "articles/single_story", story: story %>
+    <% cache("article-single-story-#{story}", expires_in: 3.minutes) do %>
+      <%= render "articles/single_story", story: story %>
+    <% end %>
   <% end %>
 <% end %>
 <% if @stories.size > 1 %>