--- a/app/views/stories/_main_stories_feed.html.erb
+++ b/app/views/stories/_main_stories_feed.html.erb
@@ -55,7 +55,7 @@
     <% if !user_signed_in? && i == 4 %>
       <%= render "stories/sign_in_invitation" %>
     <% end %>
-    <% cache("article-single-story-#{story}", expires_in: 3.minutes) do %>
+    <% cache(story, expires_in: 10.minutes) do %>
       <%= render "articles/single_story", story: story %>
     <% end %>
   <% end %>