--- a/config/application.rb
+++ b/config/application.rb
@@ -54,5 +54,7 @@ module PracticalDeveloper
       end
       ReservedWords.all = [ReservedWords::BASE_WORDS + top_routes].flatten.compact.uniq
     end
+
+    config.skylight.environments += ['development']
   end
 end