--- a/optimizations/ab_before
+++ b/optimizations/ab_before
@@ -1,0 +1,85 @@
+# ab -n 10 -c 1 127.0.0.1:3000/
+Concurrency 1
+Requests 10
+
+Server Software:
+Server Hostname:        127.0.0.1
+Server Port:            3000
+
+Document Path:          /
+Document Length:        29 bytes
+
+Concurrency Level:      1
+Time taken for tests:   0.014 seconds
+Complete requests:      10
+Failed requests:        0
+Non-2xx responses:      10
+Total transferred:      2150 bytes
+HTML transferred:       290 bytes
+Requests per second:    693.14 [#/sec] (mean)
+Time per request:       1.443 [ms] (mean)
+Time per request:       1.443 [ms] (mean, across all concurrent requests)
+Transfer rate:          145.53 [Kbytes/sec] received
+
+Connection Times (ms)
+              min  mean[+/-sd] median   max
+Connect:        0    0   0.0      0       0
+Processing:     0    1   3.3      0      11
+Waiting:        0    1   3.2      0      10
+Total:          0    1   3.3      0      11
+
+Percentage of the requests served within a certain time (ms)
+  50%      0
+  66%      0
+  75%      0
+  80%      0
+  90%     11
+  95%     11
+  98%     11
+  99%     11
+ 100%     11 (longest request)
+
+
+
+
+# ab -n 10 -c 5 127.0.0.1:3000/
+Concurrency 5
+Requests 10
+
+
+Server Software:
+Server Hostname:        127.0.0.1
+Server Port:            3000
+
+Document Path:          /
+Document Length:        29 bytes
+
+Concurrency Level:      5
+Time taken for tests:   0.010 seconds
+Complete requests:      10
+Failed requests:        0
+Non-2xx responses:      10
+Total transferred:      2150 bytes
+HTML transferred:       290 bytes
+Requests per second:    1041.34 [#/sec] (mean)
+Time per request:       4.801 [ms] (mean)
+Time per request:       0.960 [ms] (mean, across all concurrent requests)
+Transfer rate:          218.64 [Kbytes/sec] received
+
+Connection Times (ms)
+              min  mean[+/-sd] median   max
+Connect:        0    0   0.3      0       1
+Processing:     1    2   1.2      2       5
+Waiting:        1    2   1.1      2       4
+Total:          1    2   1.1      2       5
+
+Percentage of the requests served within a certain time (ms)
+  50%      2
+  66%      2
+  75%      3
+  80%      3
+  90%      5
+  95%      5
+  98%      5
+  99%      5
+ 100%      5 (longest request)