--- a/optimizations/ab_before
+++ b/optimizations/ab_before
@@ -4,82 +4,76 @@ Requests 10
 
 Server Software:
 Server Hostname:        127.0.0.1
-Server Port:            3000
+Server Port:            3001
 
 Document Path:          /
-Document Length:        29 bytes
+Document Length:        175800 bytes
 
 Concurrency Level:      1
-Time taken for tests:   0.014 seconds
+Time taken for tests:   7.000 seconds
 Complete requests:      10
 Failed requests:        0
-Non-2xx responses:      10
-Total transferred:      2150 bytes
-HTML transferred:       290 bytes
-Requests per second:    693.14 [#/sec] (mean)
-Time per request:       1.443 [ms] (mean)
-Time per request:       1.443 [ms] (mean, across all concurrent requests)
-Transfer rate:          145.53 [Kbytes/sec] received
+Total transferred:      1763730 bytes
+HTML transferred:       1758000 bytes
+Requests per second:    1.43 [#/sec] (mean)
+Time per request:       700.002 [ms] (mean)
+Time per request:       700.002 [ms] (mean, across all concurrent requests)
+Transfer rate:          246.06 [Kbytes/sec] received
 
 Connection Times (ms)
               min  mean[+/-sd] median   max
-Connect:        0    0   0.0      0       0
-Processing:     0    1   3.3      0      11
-Waiting:        0    1   3.2      0      10
-Total:          0    1   3.3      0      11
+Connect:        0    0   0.1      0       0
+Processing:   646  700  60.6    679     801
+Waiting:      645  700  60.6    679     801
+Total:        646  700  60.7    679     801
 
 Percentage of the requests served within a certain time (ms)
-  50%      0
-  66%      0
-  75%      0
-  80%      0
-  90%     11
-  95%     11
-  98%     11
-  99%     11
- 100%     11 (longest request)
-
-
-
+  50%    679
+  66%    709
+  75%    740
+  80%    800
+  90%    801
+  95%    801
+  98%    801
+  99%    801
+ 100%    801 (longest request)
 
 # ab -n 10 -c 5 127.0.0.1:3000/
 Concurrency 5
 Requests 10
 
-
 Server Software:
 Server Hostname:        127.0.0.1
-Server Port:            3000
+Server Port:            3001
 
 Document Path:          /
-Document Length:        29 bytes
+Document Length:        175800 bytes
 
 Concurrency Level:      5
-Time taken for tests:   0.010 seconds
+Time taken for tests:   4.192 seconds
 Complete requests:      10
 Failed requests:        0
-Non-2xx responses:      10
-Total transferred:      2150 bytes
-HTML transferred:       290 bytes
-Requests per second:    1041.34 [#/sec] (mean)
-Time per request:       4.801 [ms] (mean)
-Time per request:       0.960 [ms] (mean, across all concurrent requests)
-Transfer rate:          218.64 [Kbytes/sec] received
+Total transferred:      1763730 bytes
+HTML transferred:       1758000 bytes
+Requests per second:    2.39 [#/sec] (mean)
+Time per request:       2095.846 [ms] (mean)
+Time per request:       419.169 [ms] (mean, across all concurrent requests)
+Transfer rate:          410.91 [Kbytes/sec] received
 
 Connection Times (ms)
               min  mean[+/-sd] median   max
-Connect:        0    0   0.3      0       1
-Processing:     1    2   1.2      2       5
-Waiting:        1    2   1.1      2       4
-Total:          1    2   1.1      2       5
+Connect:        0    0   0.0      0       0
+Processing:   823 1532 408.9   1442    2068
+Waiting:      822 1531 408.4   1441    2067
+Total:        823 1532 408.9   1442    2068
 
 Percentage of the requests served within a certain time (ms)
-  50%      2
-  66%      2
-  75%      3
-  80%      3
-  90%      5
-  95%      5
-  98%      5
-  99%      5
- 100%      5 (longest request)
+  50%   1442
+  66%   1448
+  75%   2056
+  80%   2067
+  90%   2068
+  95%   2068
+  98%   2068
+  99%   2068
+ 100%   2068 (longest request)