… Gemfile.lock
… optimizations/new_relic_before_after_database_1.png
… optimizations/new_relic_before_after_transactions_1.png
… optimizations/new_relic_before_after_transactions_1_1.png
… optimizations/new_relic_before_after_transactions_2.png
… optimizations/new_relic_before_after_transactions_2_2.png
… optimizations/new_relic_before_after_transactions_2_3.png
… optimizations/new_relic_before_and_after_overview_1.png
… optimizations/new_relic_in_between_optimization.png
… optimizations/new_relic_transactions_1_before.png
… optimizations/newrelic_transactions_2_before.png
… optimizations/rails_panel_stories_controller_1_before.png
… optimizations/skylight_after_2.png
… optimizations/skylight_in_between_optimization.png
… optimizations/skylight_response_time_distribution_1_before.png
… optimizations/skylight_stories_controller_1_before.png