## Description
We encountered problems where our stories were loading slowly.
We determined that our agony point was located in `StoriesController#index`.

From Skylight report we can see that response time distribution is around 648 ms.
![skylight_stories_controller_1_before](https://user-images.githubusercontent.com/16690352/56581800-c6268a80-65a3-11e9-8315-ff5b9f6ca830.png)

Having chosen one of the biggest number numbers request colum, we can see that a lot of time on rendering views apart from `layouts/application.html.erb` is allocated on rendering such partials as `_main_stories_feed.html.erb` and particularly `articles/_single_story.html.erb`

![skylight_response_time_distribution_1_before](https://user-images.githubusercontent.com/16690352/56581941-143b8e00-65a4-11e9-8643-1f8986715686.png)

In order to get some data I `siege` 3 times to "fake" GET requests with the following command:
```
  siege -c 20 -t200s -f tmp/siege_urls.txt
```

## Optimization
It made sense to set cache on rendering partial `single_story` with an expiration date, so that other data depending on it (likes, comments) could be renewed.

## Benchmark
In order to measure our results before/after I used ab test.
It will allow to get fast feedback-loop.

#### Comparing results with concurrency 1 we can see the following:
* Time per request was reduced from:
```
Time per request:       700.002 [ms] (mean)
Time per request:       700.002 [ms] (mean,
```
to
```
Time per request:       567.175 [ms] (mean)
Time per request:       567.175 [ms] (mean)
```
* Requests per second increased from 1.43 to 1.76
* Transfer rate increase from 246.06 [Kbytes/sec] to 304.49 [Kbytes/sec]

#### Comparing results with concurrency 5 we can observe same positive tendency:
* Time per request was reduced from 
```
Time per request:       2095.846 [ms] (mean)
Time per request:       419.169 [ms] (mean, across all concurrent requests)
```
to
```
Time per request:       1649.773 [ms] (mean)
Time per request:       329.954 [ms] (mean, across all concurrent requests)
```
* Requests per second increased from 2.39 to 3.03
* Transfer rate increase from 410.91 [Kbytes/sec] to 523.40 [Kbytes/sec]
* Percentage of the requests served within a certain time (ms) has a big difference 
up until 80% where difference is around 400 ms:
It changed from:
```
Percentage of the requests served within a certain time (ms)
  50%   1442
  66%   1448
  75%   2056
  80%   2067
  90%   2068
  95%   2068
  98%   2068
  99%   2068
 100%   2068 (longest request)
```
to
```
Percentage of the requests served within a certain time (ms)
  50%   1038
  66%   1053
  75%   1491
  80%   1674
  90%   1925
  95%   1925
  98%   1925
  99%   1925
 100%   1925 (longest request)
```

## Full results are shown below:

### Before
```
# ab -n 10 -c 1 127.0.0.1:3001/
Concurrency 1
Requests 10

Server Software:
Server Hostname:        127.0.0.1
Server Port:            3001

Document Path:          /
Document Length:        175800 bytes

Concurrency Level:      1
Time taken for tests:   7.000 seconds
Complete requests:      10
Failed requests:        0
Total transferred:      1763730 bytes
HTML transferred:       1758000 bytes
Requests per second:    1.43 [#/sec] (mean)
Time per request:       700.002 [ms] (mean)
Time per request:       700.002 [ms] (mean, across all concurrent requests)
Transfer rate:          246.06 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       0
Processing:   646  700  60.6    679     801
Waiting:      645  700  60.6    679     801
Total:        646  700  60.7    679     801

Percentage of the requests served within a certain time (ms)
  50%    679
  66%    709
  75%    740
  80%    800
  90%    801
  95%    801
  98%    801
  99%    801
 100%    801 (longest request)

# ab -n 10 -c 5 127.0.0.1:3001/
Concurrency 5
Requests 10

Server Software:
Server Hostname:        127.0.0.1
Server Port:            3001

Document Path:          /
Document Length:        175800 bytes

Concurrency Level:      5
Time taken for tests:   4.192 seconds
Complete requests:      10
Failed requests:        0
Total transferred:      1763730 bytes
HTML transferred:       1758000 bytes
Requests per second:    2.39 [#/sec] (mean)
Time per request:       2095.846 [ms] (mean)
Time per request:       419.169 [ms] (mean, across all concurrent requests)
Transfer rate:          410.91 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   823 1532 408.9   1442    2068
Waiting:      822 1531 408.4   1441    2067
Total:        823 1532 408.9   1442    2068

Percentage of the requests served within a certain time (ms)
  50%   1442
  66%   1448
  75%   2056
  80%   2067
  90%   2068
  95%   2068
  98%   2068
  99%   2068
 100%   2068 (longest request)
```

### After
```# ab -n 10 -c 1 127.0.0.1:3001/
Concurrency 1
Requests 10

Server Software:
Server Hostname:        127.0.0.1
Server Port:            3001

Document Path:          /
Document Length:        176270 bytes

Concurrency Level:      1
Time taken for tests:   5.672 seconds
Complete requests:      10
Failed requests:        0
Total transferred:      1768430 bytes
HTML transferred:       1762700 bytes
Requests per second:    1.76 [#/sec] (mean)
Time per request:       567.175 [ms] (mean)
Time per request:       567.175 [ms] (mean, across all concurrent requests)
Transfer rate:          304.49 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   434  567  95.2    562     717
Waiting:      434  567  95.2    562     717
Total:        435  567  95.2    562     717

Percentage of the requests served within a certain time (ms)
  50%    562
  66%    624
  75%    629
  80%    681
  90%    717
  95%    717
  98%    717
  99%    717
 100%    717 (longest request)

# ab -n 10 -c 5 127.0.0.1:3001/
Concurrency 5
Requests 10

Server Software:
Server Hostname:        127.0.0.1
Server Port:            3001

Document Path:          /
Document Length:        176270 bytes

Concurrency Level:      5
Time taken for tests:   3.300 seconds
Complete requests:      10
Failed requests:        0
Total transferred:      1768430 bytes
HTML transferred:       1762700 bytes
Requests per second:    3.03 [#/sec] (mean)
Time per request:       1649.773 [ms] (mean)
Time per request:       329.954 [ms] (mean, across all concurrent requests)
Transfer rate:          523.40 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   636 1177 391.6   1037    1925
Waiting:      636 1176 390.7   1037    1924
Total:        636 1177 391.5   1038    1925

Percentage of the requests served within a certain time (ms)
  50%   1038
  66%   1053
  75%   1491
  80%   1674
  90%   1925
  95%   1925
  98%   1925
  99%   1925
 100%   1925 (longest request)
```

## Conclusion
We can see that caching strategy helped us to optimize this page a bit.

### Note
During this assignment I used siege with and without `cache` multiple times to check how graphs would look over time for comparison purposes

For NewRelic 
- Not optimized time:
- 6:30pm-7:00pm
- 7:30pm-8:15pm

- Optimized time
- 7:00 - 7:30pm
- 8:15pm - and later

From Newrelic report how our app behaves with and withouth caching:

<img width="1680" alt="new_relic_before_and_after_overview_1" src="https://user-images.githubusercontent.com/16690352/56584105-09372c80-65a9-11e9-9da1-d372a232a488.png">
<img width="1680" alt="new_relic_before_and_after_overview_1" src="https://user-images.githubusercontent.com/16690352/56584407-b0b45f00-65a9-11e9-85d0-56746fc20327.png">
<img width="1665" alt="new_relic_before_after_transactions_1" src="https://user-images.githubusercontent.com/16690352/56584519-e3f6ee00-65a9-11e9-8304-ff31460522ee.png">
<img width="1656" alt="new_relic_before_after_transactions_2_2" src="https://user-images.githubusercontent.com/16690352/56584520-e48f8480-65a9-11e9-8241-b3a1373ffabb.png">
<img width="1670" alt="new_relic_before_after_transactions_2_3" src="https://user-images.githubusercontent.com/16690352/56584522-e48f8480-65a9-11e9-94d9-c8358681251f.png">
<img width="1673" alt="new_relic_before_after_transactions_2" src="https://user-images.githubusercontent.com/16690352/56584523-e48f8480-65a9-11e9-9d62-ed7f3b5d6df4.png">
<img width="1680" alt="new_relic_before_and_after_overview_1" src="https://user-images.githubusercontent.com/16690352/56584524-e48f8480-65a9-11e9-88d6-18f6796db6c3.png">



Skyligh:
- Not optimized 
- 9:00pm-9:30pm
- 10:25pm-10:55pm

- Optimized
- 9:30pm-10:25pm
- 11:05pm - after
<img width="1668" alt="skylight_after_2" src="https://user-images.githubusercontent.com/16690352/56584804-8adb8a00-65aa-11e9-97ef-06c6663e7927.png">



We can see that response time distribution is around 363 ms comparing to 648 ms in the beggining.
<img width="1679" alt="skylight_after_2_2" src="https://user-images.githubusercontent.com/16690352/56584869-ae9ed000-65aa-11e9-888b-0a0a77f54419.png">